Add files via upload
This program can change the specific value in images. When using other library or software, we can only change the setting of no-data value but we cannot really change the value in the image. While when using ENVI deep learning module, it seems that the no-data value in the images should be the same as training data. Thus, it is necessary to change the real value in images instead of the setting of no-data value.
I think some bug exists due to the threading things, which I am not familiar with. Hope I can fix it in next commit.