Add files via upload
This program can create a image with existing image and complicated band math. The new image is composed by the existing image and the result of band math. I create this for testing whether the AI can achieve better result with specific features. For now, the new band is NDVI.